pyproject.toml
[project]
name = "process-template"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "automation-server-client @ git+https://github.com/odense-rpa/automation-server-client.git",
    "mbu-dev-shared-components>=3.0.0",
    "mbu-rpa-core",
    "pillow",
    "ruff"
]

[tool.uv.sources]
automation-server-client = { git = "https://github.com/odense-rpa/automation-server-client.git", tag = "v0.2.0" }

[tool.setuptools]
py-modules = []

[tool.ruff]
target-version = "py313"
exclude = [
  ".git",
  ".venv",
  "venv",
  "__pycache__",
  "build",
  "dist",
  "uv.lock",
  ".env",
]

[tool.ruff.lint]
ignore = ["E501", "E251"]
select = [
  "F",    # Pyflakes
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "I",    # isort (import sorting)
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "TC",   # flake8-type-checking
  "ARG",  # flake8-unused-arguments
  "DTZ",  # enforce timezone-aware datetimes
  "SIM",  # flake8-simplify
  "PL",   # pylint (subset)
]

[tool.ruff.lint.isort]
# Keep imports tidy and deterministic
# known-first-party = ["your_package_name"]

[tool.ruff.format]
# Use the formatter; CI step already checks it with `ruff format --check .`

[tool.pyright]
include = ["."]
exclude = [
  "venv",
  ".venv",
  "build",
  "dist",
  "__pycache__"
]
pythonVersion = "3.13"
typeCheckingMode = "basic"   # or "basic"
reportMissingTypeStubs = true
